Oops, IPC coding requires lots of care, tracking instruction stream is really torching me. But at last I found where the bug is, now messages could be exchanged between different Processes.

+++ include/type.h
@@ -1,10 +1,10 @@
 #ifndef	_LAVENDER_TYPE_H_
 #define	_LAVENDER_TYPE_H_
 
-typedef	unsigned long	u64;
-typedef	unsigned int	u32;
-typedef unsigned short	u16;
-typedef	unsigned char	u8;
+typedef	unsigned long long	u64;
+typedef	unsigned int		u32;
+typedef unsigned short		u16;
+typedef	unsigned char		u8;
 
 typedef	void	(* int_handler)();
 typedef	void	(* irq_handler)(int irq);
